pyproject.toml
[project]
name = "ai-image-classifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "keras>=3.13.0",
    "matplotlib>=3.10.8",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "redis>=7.1.0",
    "reportlab>=4.4.9",
    "resend>=2.19.0",
    "rq>=2.6.1",
    "scikit-learn>=1.8.0",
    "supabase>=2.27.2",
    "supabase-auth>=2.27.2",
    "tensorflow>=2.20.0",
    "tqdm>=4.67.1",
    "uvicorn>=0.40.0",
]
